MNT: Rename table_widgets.py -> color_button.py
